thermal: rcar_gen3_thermal: Use DIV_ROUND_CLOSEST correctly as its description
As description for DIV_ROUND_CLOSEST in file include/linux/kernel.h.
  "Result is undefined for negative divisors
   if the dividend variable type is unsigned
   and for negative dividends
   if the divisor variable type is unsigned."

In current code, the FIXPT_DIV uses DIV_ROUND_CLOSEST
but has not checked sign of divisor before using.
It makes undefined temperature value in case
the value is negative.

This patch fixes to satisfy DIV_ROUND_CLOSEST description
and fix bug too.

+++ drivers/thermal/rcar_gen3_thermal.c
@@ -216,10 +216,12 @@ static int rcar_gen3_thermal_round(int temp)
 static int rcar_gen3_thermal_convert_temp(struct rcar_gen3_thermal_tsc *tsc)
 {
 	int mcelsius = 0, val1, val2;
-	u32 reg;
+	long reg;
 
 	if (is_ths_typeA) {
 		/* Read register and convert to mili Celsius */
+
+		/* Read register and convert value to signed variable type */
 		reg = rcar_gen3_thermal_read(tsc, REG_GEN3_TEMP) & CTEMP_MASK;
 
 		val1 = FIXPT_DIV(FIXPT_INT(reg) - tsc->coef.b1, tsc->coef.a1);
